pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neuro-magic-courses"
version = "0.1.0"
description = "Нейромагический MVP на Quart с фронтендом Vite/React."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib==1.3.2",
    "httpx==0.27.0",
    "itsdangerous==2.2.0",
    "loguru==0.7.2",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "quart==0.19.6"
]

[project.optional-dependencies]
dev = [
    "mypy==1.11.2",
    "pytest==8.3.2",
    "pytest-asyncio==0.23.8",
    "ruff==0.6.7"
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",
    "F",
    "B",
    "I",
    "UP"
]

